added all updates

+++ src/components/ManageOrders/ManageOrder.js
@@ -4,31 +4,31 @@ import Button from 'react-bootstrap/Button'
 
 const ManageOrder = () => {
     const [orders, setOrders] = useState([])
-    useEffect(()=>{
-        fetch('http://localhost:5000/TourService/booking')
-        .then(res =>res.json())
-        .then( data => setOrders(data))
-    },[])
+    useEffect(() => {
+        fetch('https://whispering-dusk-80653.herokuapp.com/TourService/booking')
+            .then(res => res.json())
+            .then(data => setOrders(data))
+    }, [])
 
-    const handleDelete = id =>{
-        fetch(`http://localhost:5000/TourService/booking/${id}`, {
-            method : 'DELETE'
-        })
-        .then(res =>res.json())
-        .then(data =>{
-                if(data.acknowledged){
-                    alert('Delete Booking Successfully')
-                    const restBooking = orders.filter(order=> order._id !== id)
-                    setOrders(restBooking)
-                    
-                }
-        })
+    const handleDelete = id => {
+        const proceed = window.confirm('Are you sure, you want to delete?');
+        if (proceed) {
+            fetch(`https://whispering-dusk-80653.herokuapp.com/TourService/booking/${id}`, {
+                method: 'DELETE'
+            })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.acknowledged) {
+                        alert('Delete Booking Successfully')
+                        const restBooking = orders.filter(order => order._id !== id)
+                        setOrders(restBooking)
+                    }
+                })
+        }
     }
-    
-
     return (
-        <div className = 'my-5'>
-            <h2 className = 'my-5 text-center fw-bold custom-text-primary'>Manage All Booking Orders</h2>
+        <div className='my-5'>
+            <h2 className='my-5 text-center fw-bold custom-text-primary'>Manage All Booking Orders</h2>
             <Table responsive="sm" >
                 <thead >
                     <tr>
@@ -42,14 +42,14 @@ const ManageOrder = () => {
                 </thead>
                 <tbody>
                     {
-                        orders.map(order =><tr>
+                        orders.map(order => <tr>
                             <td>{order._id}</td>
                             <td>{order?.name}</td>
                             <td>{order?.email}</td>
                             <td>{order?.Country}</td>
-                            <td><Button onClick = {()=> handleDelete(order._id)} variant="danger">Cancel</Button></td>
+                            <td><Button onClick={() => handleDelete(order._id)} variant="danger">Cancel</Button></td>
                             <td>{order.status}</td>
-                            
+
                         </tr>)
                     }
                 </tbody>
